# products/src/main/java/com/ucc/crud_service/model/Product.java
package com.ucc.crud_service.model;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotBlank;
import lombok.*;

@Entity
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class Product {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @NotBlank(message = "El campo Descripcion es obligatorio")
    private String description;

    @DecimalMin(value = "0.00", message = "El precio no puede ser cero")
    private double price;

    private Long Stock;
}